[CCUBE-242][KT] Move helper functions to an exported namespace
